changes for release (#2270)
