+++ packages/salesforcedx-vscode-core/src/commands/forceSourceRetrieveSourcePath.ts
@@ -32,10 +32,11 @@ import {
   SfdxWorkspaceChecker
 } from './util';
 import { LibraryCommandletExecutor } from './util/libraryCommandlet';
+import { useBetaDeployRetrieve } from './util/useBetaDeployRetrieve';
 
 export class ForceSourceRetrieveSourcePathExecutor extends SfdxCommandletExecutor<
   string
-> {
+  > {
   public build(sourcePath: string): Command {
     return new SfdxCommandBuilder()
       .withDescription(nls.localize('force_source_retrieve_text'))
@@ -104,7 +105,7 @@ export async function forceSourceRetrieveSourcePath(explorerPath: vscode.Uri) {
   const commandlet = new SfdxCommandlet(
     new SfdxWorkspaceChecker(),
     new FilePathGatherer(explorerPath),
-    useBetaRetrieve(explorerPath)
+    useBetaDeployRetrieve([explorerPath])
       ? new LibraryRetrieveSourcePathExecutor()
       : new ForceSourceRetrieveSourcePathExecutor(),
     new SourcePathChecker()
@@ -112,36 +113,9 @@ export async function forceSourceRetrieveSourcePath(explorerPath: vscode.Uri) {
   await commandlet.run();
 }
 
-// this supported types logic is temporary until we have a way of generating the metadata type from the path
-// once we have the metadata type we can check to see if it is a toolingsupportedtype from that util
-export function useBetaRetrieve(explorerPath: vscode.Uri): boolean {
-  const filePath = explorerPath.fsPath;
-  const betaDeployRetrieve = sfdxCoreSettings.getBetaDeployRetrieve();
-  const registry = new RegistryAccess();
-  const component = registry.getComponentsFromPath(filePath)[0];
-  const typeName = component.type.name;
-  const {
-    auradefinitionbundle,
-    lightningcomponentbundle,
-    apexclass,
-    apexcomponent,
-    apexpage,
-    apextrigger
-  } = registryData.types;
-
-  const supportedType =
-    typeName === auradefinitionbundle.name ||
-    typeName === lightningcomponentbundle.name ||
-    typeName === apexclass.name ||
-    typeName === apexcomponent.name ||
-    typeName === apexpage.name ||
-    typeName === apextrigger.name;
-  return betaDeployRetrieve && supportedType;
-}
-
 export class LibraryRetrieveSourcePathExecutor extends LibraryCommandletExecutor<
   string
-> {
+  > {
   public async execute(response: ContinueResponse<string>): Promise<void> {
     this.setStartTime();
