+++ packages/salesforcedx-vscode-core/test/vscode-integration/commands/forceSourceRetrieveSourcePath.test.ts
@@ -9,21 +9,16 @@ import {
   CancelResponse,
   ContinueResponse
 } from '@salesforce/salesforcedx-utils-vscode/out/src/types/index';
-import { RegistryAccess } from '@salesforce/source-deploy-retrieve';
-import { MetadataComponent } from '@salesforce/source-deploy-retrieve/lib/types';
 import { expect } from 'chai';
 import * as path from 'path';
 import { createSandbox, SinonSandbox, SinonStub } from 'sinon';
-import { Uri } from 'vscode';
 import { channelService } from '../../../src/channels';
 import {
   ForceSourceRetrieveSourcePathExecutor,
-  SourcePathChecker,
-  useBetaRetrieve
+  SourcePathChecker
 } from '../../../src/commands/forceSourceRetrieveSourcePath';
 import { nls } from '../../../src/messages';
 import { notificationService } from '../../../src/notifications';
-import { SfdxCoreSettings } from '../../../src/settings/sfdxCoreSettings';
 import { SfdxPackageDirectories } from '../../../src/sfdxProject';
 import { getRootWorkspacePath } from '../../../src/util';
 
@@ -114,227 +109,5 @@ describe('SourcePathChecker', () => {
     expect(showErrorMessageSpy.getCall(0).args[0]).to.equal(errorMessage);
     expect(cancelResponse.type).to.equal('CANCEL');
     isInPackageDirectoryStub.restore();
-  });
-});
-
-describe('Force Source Retrieve with Sourcepath Beta', () => {
-  let sandboxStub: SinonSandbox;
-  let mockRegistry: SinonStub;
-  const apexClassMDComponent: MetadataComponent = {
-    type: {
-      name: 'ApexClass',
-      directoryName: 'classes',
-      inFolder: false,
-      suffix: 'cls'
-    },
-    fullName: 'myTestClass',
-    xml: path.join('file', 'path', 'classes', 'myTestClass.cls-meta.xml'),
-    sources: [path.join('file', 'path', 'classes', 'myTestClass.cls')]
-  };
-  const apexTriggerMDComponent: MetadataComponent = {
-    type: {
-      name: 'ApexTrigger',
-      directoryName: 'triggers',
-      inFolder: false,
-      suffix: 'trigger'
-    },
-    fullName: 'accTrigger',
-    xml: path.join('file', 'path', 'triggers', 'accTrigger.trigger-meta.xml'),
-    sources: [path.join('file', 'path', 'triggers', 'accTrigger.trigger')]
-  };
-  const pageMDComponent: MetadataComponent = {
-    type: {
-      name: 'ApexPage',
-      directoryName: 'pages',
-      inFolder: false,
-      suffix: 'page'
-    },
-    fullName: 'myPage',
-    xml: path.join('file', 'path', 'pages', 'myPage.page-meta.xml'),
-    sources: [path.join('file', 'path', 'pages', 'myPage.page')]
-  };
-  const vfComponentMDComponent: MetadataComponent = {
-    type: {
-      name: 'ApexComponent',
-      directoryName: 'components',
-      inFolder: false,
-      suffix: 'component'
-    },
-    fullName: 'myPage',
-    xml: path.join('file', 'path', 'components', 'VFCmp.component-meta.xml'),
-    sources: [path.join('file', 'path', 'components', 'VFCmp.component')]
-  };
-  const auraMDComponent: MetadataComponent = {
-    type: {
-      name: 'AuraDefinitionBundle',
-      directoryName: 'aura',
-      inFolder: false
-    },
-    fullName: 'testApp',
-    xml: path.join('file', 'path', 'aura', 'testApp.app-meta.xml'),
-    sources: [path.join('file', 'path', 'aura', 'testApp.app')]
-  };
-  const lwcMDComponent: MetadataComponent = {
-    type: {
-      name: 'LightningComponentBundle',
-      directoryName: 'lwc',
-      inFolder: false
-    },
-    fullName: 'testCmp',
-    xml: path.join('file', 'path', 'lwc', 'testCmp', 'testCmp.js-meta.xml'),
-    sources: [path.join('file', 'path', 'lwc', 'testCmp', 'testCmp.js')]
-  };
-
-  beforeEach(() => {
-    sandboxStub = createSandbox();
-    mockRegistry = sandboxStub.stub(
-      RegistryAccess.prototype,
-      'getComponentsFromPath'
-    );
-  });
-
-  afterEach(() => {
-    sandboxStub.restore();
-  });
-
-  it('Should return false for URI not part of the beta when the beta configuration is enabled', () => {
-    sandboxStub
-      .stub(SfdxCoreSettings.prototype, 'getBetaDeployRetrieve')
-      .returns(true);
-    mockRegistry.returns([
-      {
-        type: {
-          name: 'Layout',
-          directoryName: 'layouts',
-          inFolder: false
-        },
-        fullName: 'Obj Layout',
-        xml: path.join('file', 'path', 'layouts', 'Obj Layout.layout-meta.xml'),
-        sources: [path.join('file', 'path', 'layouts', 'Obj Layout.layout')]
-      }
-    ]);
-    const uriOne = Uri.parse('file:///bar.html');
-    const fileProcessing = useBetaRetrieve(uriOne);
-    expect(fileProcessing).to.equal(false);
-  });
-
-  it('Should return true for ApexClass URI when beta configuration is enabled', () => {
-    sandboxStub
-      .stub(SfdxCoreSettings.prototype, 'getBetaDeployRetrieve')
-      .returns(true);
-    mockRegistry.returns([apexClassMDComponent]);
-    const uriOne = Uri.parse('file:///file/path/classes/myTestClass.cls');
-    const apexClassProcessing = useBetaRetrieve(uriOne);
-    expect(apexClassProcessing).to.equal(true);
-  });
-
-  it('Should return false for ApexClass URI when beta configuration is disabled', () => {
-    sandboxStub
-      .stub(SfdxCoreSettings.prototype, 'getBetaDeployRetrieve')
-      .returns(false);
-    mockRegistry.returns([apexClassMDComponent]);
-    const uriOne = Uri.parse('file:///file/path/classes/myTestClass.cls');
-    const apexClassProcessing = useBetaRetrieve(uriOne);
-    expect(apexClassProcessing).to.equal(false);
-  });
-
-  it('Should return true for ApexTrigger URI when beta configuration is enabled', () => {
-    sandboxStub
-      .stub(SfdxCoreSettings.prototype, 'getBetaDeployRetrieve')
-      .returns(true);
-    mockRegistry.returns([apexTriggerMDComponent]);
-    const uriOne = Uri.parse('file:///file/path/triggers/accTrigger.trigger');
-    const triggerProcessing = useBetaRetrieve(uriOne);
-    expect(triggerProcessing).to.equal(true);
-  });
-
-  it('Should return false for ApexTrigger URI when beta configuration is disabled', () => {
-    sandboxStub
-      .stub(SfdxCoreSettings.prototype, 'getBetaDeployRetrieve')
-      .returns(false);
-    mockRegistry.returns([apexTriggerMDComponent]);
-    const uriOne = Uri.parse('file:///file/path/triggers/accTrigger.trigger');
-    const triggerProcessing = useBetaRetrieve(uriOne);
-    expect(triggerProcessing).to.equal(false);
-  });
-
-  it('Should return true for VF Page URI when beta configuration is enabled', () => {
-    sandboxStub
-      .stub(SfdxCoreSettings.prototype, 'getBetaDeployRetrieve')
-      .returns(true);
-    mockRegistry.returns([pageMDComponent]);
-    const uriOne = Uri.parse('file:///file/path/pages/myPage.page');
-    const pageProcessing = useBetaRetrieve(uriOne);
-    expect(pageProcessing).to.equal(true);
-  });
-
-  it('Should return false for VF Page URI when beta configuration is disabled', () => {
-    sandboxStub
-      .stub(SfdxCoreSettings.prototype, 'getBetaDeployRetrieve')
-      .returns(false);
-    mockRegistry.returns([pageMDComponent]);
-    const uriOne = Uri.parse('file:///file/path/pages/myPage.page');
-    const pageProcessing = useBetaRetrieve(uriOne);
-    expect(pageProcessing).to.equal(false);
-  });
-
-  it('Should return true for VF Component URI when beta configuration is enabled', () => {
-    sandboxStub
-      .stub(SfdxCoreSettings.prototype, 'getBetaDeployRetrieve')
-      .returns(true);
-    mockRegistry.returns([vfComponentMDComponent]);
-    const uriOne = Uri.parse('file:///file/path/components/VFCmp.component');
-    const cmpProcessing = useBetaRetrieve(uriOne);
-    expect(cmpProcessing).to.equal(true);
-  });
-
-  it('Should return false for VF Component URI when beta configuration is disabled', () => {
-    sandboxStub
-      .stub(SfdxCoreSettings.prototype, 'getBetaDeployRetrieve')
-      .returns(false);
-    mockRegistry.returns([vfComponentMDComponent]);
-    const uriOne = Uri.parse('file:///file/path/components/VFCmp.component');
-    const cmpProcessing = useBetaRetrieve(uriOne);
-    expect(cmpProcessing).to.equal(false);
-  });
-
-  it('Should return true for Aura Component URI when beta configuration is enabled', () => {
-    sandboxStub
-      .stub(SfdxCoreSettings.prototype, 'getBetaDeployRetrieve')
-      .returns(true);
-    mockRegistry.returns([auraMDComponent]);
-    const uriOne = Uri.parse('file:///file/path/aura/testApp.app');
-    const cmpProcessing = useBetaRetrieve(uriOne);
-    expect(cmpProcessing).to.equal(true);
-  });
-
-  it('Should return false for Aura Component URI when beta configuration is disabled', () => {
-    sandboxStub
-      .stub(SfdxCoreSettings.prototype, 'getBetaDeployRetrieve')
-      .returns(false);
-    mockRegistry.returns([auraMDComponent]);
-    const uriOne = Uri.parse('file:///file/path/aura/testApp.app');
-    const cmpProcessing = useBetaRetrieve(uriOne);
-    expect(cmpProcessing).to.equal(false);
-  });
-
-  it('Should return true for LWC Component URI when beta configuration is enabled', () => {
-    sandboxStub
-      .stub(SfdxCoreSettings.prototype, 'getBetaDeployRetrieve')
-      .returns(true);
-    mockRegistry.returns([lwcMDComponent]);
-    const uriOne = Uri.parse('file:///file/path/lwc/testCmp/testCmp.js');
-    const cmpProcessing = useBetaRetrieve(uriOne);
-    expect(cmpProcessing).to.equal(true);
-  });
-
-  it('Should return false for LWC Component URI when beta configuration is disabled', () => {
-    sandboxStub
-      .stub(SfdxCoreSettings.prototype, 'getBetaDeployRetrieve')
-      .returns(false);
-    mockRegistry.returns([lwcMDComponent]);
-    const uriOne = Uri.parse('file:///file/path/lwc/testCmp/testCmp.js');
-    const cmpProcessing = useBetaRetrieve(uriOne);
-    expect(cmpProcessing).to.equal(false);
   });
 });
